+++ app/Http/Controllers/Escola/Projeto/EscolaProjetoController.php
@@ -13,6 +13,7 @@ use App\Projeto;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Input;
 
 class EscolaProjetoController extends Controller
@@ -81,6 +82,7 @@ class EscolaProjetoController extends Controller
     {
         try {
             $projeto = Projeto::find($id);
+            $this->authorize('show', $projeto);
             $alunos = Aluno::all()
                 ->where('projeto_id', '=', $projeto->id);
             $professores = Professor::all()
@@ -96,8 +98,10 @@ class EscolaProjetoController extends Controller
         try {
             $projeto = Projeto::find($id);
             $disciplinas = Disciplina::all();
+            $this->authorize('edit', $projeto);
             $titulo = 'Editar projeto: ' . $projeto->titulo;
             return view("escola.projeto.editar", compact('projeto', 'titulo', 'disciplinas'));
+
         } catch (\Exception $e) {
             return "ERRO: " . $e->getMessage();
         }
@@ -128,6 +132,8 @@ class EscolaProjetoController extends Controller
     public function destroy($id)
     {
         try {
+            $projeto = Projeto::find($id);
+            $this->authorize('destroy', $projeto);
             $this->projetoController->destroy($id);
         } catch (\Exception $e) {
             return "ERRO: " . $e->getMessage();
